move types needed for reading old versions to new wallet legacy module
