+++ zingolib/src/wallet/data.rs
@@ -1,24 +1,17 @@
 //! TODO: Add Mod Description Here!
-use super::traits::ToBytes;
 
 use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
-use incrementalmerkletree::frontier::CommitmentTree;
 use incrementalmerkletree::witness::IncrementalWitness;
 use incrementalmerkletree::{Hashable, Position};
 
 use json::JsonValue;
-use prost::Message;
 
 use std::convert::TryFrom;
 use std::io::{self, Read, Write};
-use zcash_client_backend::{
-    proto::compact_formats::CompactBlock, wallet::TransparentAddressMetadata,
-};
 
-use zcash_encoding::{CompactSize, Optional, Vector};
+use zcash_encoding::{CompactSize, Optional};
 
-use zcash_primitives::merkle_tree::{read_commitment_tree, write_commitment_tree};
-use zcash_primitives::{legacy::TransparentAddress, memo::MemoBytes};
+use zcash_primitives::memo::MemoBytes;
 use zcash_primitives::{memo::Memo, transaction::TxId};
 
 pub use crate::wallet::transaction_record::TransactionRecord; // TODO: is this necessary? can we import this directly where its used?
@@ -27,123 +20,6 @@ pub use crate::wallet::transaction_record::TransactionRecord; // TODO: is this n
 pub const COMMITMENT_TREE_LEVELS: u8 = 32;
 /// TODO: Add Doc Comment Here!
 pub const MAX_SHARD_LEVEL: u8 = 16;
-
-/// This type is motivated by the IPC architecture where (currently) channels traffic in
-/// `(TxId, WalletNullifier, BlockHeight, Option<u32>)`.  This enum permits a single channel
-/// type to handle nullifiers from different domains.
-/// <https://github.com/zingolabs/zingolib/issues/64>
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum PoolNullifier {
-    /// TODO: Add Doc Comment Here!
-    Sapling(sapling_crypto::Nullifier),
-    /// TODO: Add Doc Comment Here!
-    Orchard(orchard::note::Nullifier),
-}
-
-impl std::hash::Hash for PoolNullifier {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        match self {
-            PoolNullifier::Sapling(n) => {
-                state.write_u8(0);
-                n.0.hash(state);
-            }
-            PoolNullifier::Orchard(n) => {
-                state.write_u8(1);
-                n.to_bytes().hash(state);
-            }
-        }
-    }
-}
-
-/// TODO: Add Doc Comment Here!
-#[derive(Clone, Debug, PartialEq)]
-pub struct BlockData {
-    /// TODO: Add Doc Comment Here!
-    pub(crate) ecb: Vec<u8>,
-    /// TODO: Add Doc Comment Here!
-    pub height: u64,
-}
-
-impl BlockData {
-    /// TODO: Add Doc Comment Here!
-    pub fn serialized_version() -> u64 {
-        20
-    }
-
-    pub(crate) fn new_with(height: u64, hash: &[u8]) -> Self {
-        let hash = hash.iter().copied().rev().collect::<Vec<_>>();
-
-        let cb = CompactBlock {
-            hash,
-            ..Default::default()
-        };
-
-        let mut ecb = vec![];
-        cb.encode(&mut ecb).unwrap();
-
-        Self { ecb, height }
-    }
-
-    pub(crate) fn cb(&self) -> CompactBlock {
-        let b = self.ecb.clone();
-        CompactBlock::decode(&b[..]).unwrap()
-    }
-
-    pub(crate) fn hash(&self) -> String {
-        self.cb().hash().to_string()
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn read<R: Read>(mut reader: R) -> io::Result<Self> {
-        let height = reader.read_i32::<LittleEndian>()? as u64;
-
-        let mut hash_bytes = [0; 32];
-        reader.read_exact(&mut hash_bytes)?;
-        hash_bytes.reverse();
-
-        // We don't need this, but because of a quirk, the version is stored later, so we can't actually
-        // detect the version here. So we write an empty tree and read it back here
-        let tree: sapling_crypto::CommitmentTree = read_commitment_tree(&mut reader)?;
-        let _tree = if tree.size() == 0 { None } else { Some(tree) };
-
-        let version = reader.read_u64::<LittleEndian>()?;
-
-        let ecb = if version <= 11 {
-            vec![]
-        } else {
-            Vector::read(&mut reader, |r| r.read_u8())?
-        };
-
-        if ecb.is_empty() {
-            Ok(BlockData::new_with(height, &hash_bytes))
-        } else {
-            Ok(BlockData { ecb, height })
-        }
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        writer.write_i32::<LittleEndian>(self.height as i32)?;
-
-        let hash_bytes: Vec<_> = hex::decode(self.hash())
-            .unwrap()
-            .into_iter()
-            .rev()
-            .collect();
-        writer.write_all(&hash_bytes[..])?;
-
-        write_commitment_tree(
-            &CommitmentTree::<sapling_crypto::Node, 32>::empty(),
-            &mut writer,
-        )?;
-        writer.write_u64::<LittleEndian>(Self::serialized_version())?;
-
-        // Write the ecb as well
-        Vector::write(&mut writer, &self.ecb, |w, b| w.write_u8(*b))?;
-
-        Ok(())
-    }
-}
 
 /// TODO: Add Doc Comment Here!
 #[derive(Clone)]
@@ -1639,46 +1515,6 @@ impl WalletZecPriceInfo {
 
         Ok(())
     }
-}
-
-/// Generate a new rejection address,
-/// for use in a send to a TEX address.
-pub fn new_rejection_address(
-    rejection_addresses: &append_only_vec::AppendOnlyVec<(
-        TransparentAddress,
-        TransparentAddressMetadata,
-    )>,
-
-    rejection_ivk: &zcash_primitives::legacy::keys::EphemeralIvk,
-) -> Result<
-    (
-        zcash_primitives::legacy::TransparentAddress,
-        zcash_client_backend::wallet::TransparentAddressMetadata,
-    ),
-    super::error::KeyError,
-> {
-    let (rejection_address, metadata) =
-        super::keys::unified::WalletCapability::get_rejection_address_by_index(
-            rejection_ivk,
-            rejection_addresses.len() as u32,
-        )?;
-    rejection_addresses.push((rejection_address, metadata.clone()));
-    Ok((rejection_address, metadata))
-}
-
-#[test]
-fn read_write_empty_sapling_tree() {
-    let mut buffer = Vec::new();
-
-    write_commitment_tree(
-        &CommitmentTree::<sapling_crypto::Node, 32>::empty(),
-        &mut buffer,
-    )
-    .unwrap();
-    assert_eq!(
-        CommitmentTree::<sapling_crypto::Node, 32>::empty(),
-        read_commitment_tree(&mut buffer.as_slice()).unwrap()
-    )
 }
 
 #[cfg(any(test, feature = "test-elevation"))]
